fix(backend): fix some file for backend implementation
