+++ app/admin/orders/page.jsx
@@ -26,7 +26,7 @@ export default function OrdersPage() {
   useEffect(() => {
     const fetchOrders = async () => {
       try {
-        const response = await axios.get("http://127.0.0.1:5000/api/v1/orders");
+        const response = await axios.get("http://localhost:8002/api/v1/orders");
         setOrders(response.data);
         console.log(response.data);
       } catch (err) {
